# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lead Tracker API",
    "description": "Local business prospecting: search, audit, score, call, export.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/export": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "Export Csv",
        "operationId": "export_csv_api_export_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Priority"
              },
              "title": "Priority"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/LeadStatus"
              },
              "title": "Status"
            }
          },
          {
            "name": "opportunity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/OpportunityType"
              },
              "title": "Opportunity"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "niche_key",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Niche Key"
            }
          },
          {
            "name": "city",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "City"
            }
          },
          {
            "name": "has_phone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Phone"
            }
          },
          {
            "name": "has_website",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Website"
            }
          },
          {
            "name": "website_problems",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Website Problems"
            }
          },
          {
            "name": "google_opportunity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Google Opportunity"
            }
          },
          {
            "name": "contacted",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contacted"
            }
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 5,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Rating"
            }
          },
          {
            "name": "min_reviews",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Reviews"
            }
          },
          {
            "name": "max_reviews",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Reviews"
            }
          },
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Score"
            }
          },
          {
            "name": "open_in_window",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Open In Window"
            }
          },
          {
            "name": "callback_due",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Callback Due"
            }
          },
          {
            "name": "assigned_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assigned To"
            }
          },
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job Id"
            }
          },
          {
            "name": "job_matched_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Job Matched Only"
            }
          },
          {
            "name": "include_suppressed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Suppressed"
            }
          },
          {
            "name": "include_demo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Include Demo"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "priority",
                "opportunity_score",
                "website_score",
                "review_count",
                "rating",
                "name",
                "created_at",
                "last_contacted_at",
                "next_callback_at"
              ],
              "type": "string",
              "default": "priority",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/preview": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Preview",
        "operationId": "import_preview_api_import_preview_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_preview_api_import_preview_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import/remap": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Remap",
        "operationId": "import_remap_api_import_remap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportCommit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/import": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Import Commit",
        "operationId": "import_commit_api_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportCommit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_preview_api_import_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_preview_api_import_preview_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportCommit": {
        "properties": {
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "mapping": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapping"
          },
          "duplicate_strategy": {
            "type": "string",
            "enum": [
              "skip",
              "fill_empty"
            ],
            "title": "Duplicate Strategy",
            "default": "skip"
          },
          "audit_websites": {
            "type": "boolean",
            "title": "Audit Websites",
            "default": true
          },
          "skip_rows": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Skip Rows"
          }
        },
        "type": "object",
        "required": [
          "batch_id"
        ],
        "title": "ImportCommit"
      },
      "ImportPreview": {
        "properties": {
          "batch_id": {
            "type": "integer",
            "title": "Batch Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns"
          },
          "mapping": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Mapping"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ImportPreviewRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "target_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Fields"
          }
        },
        "type": "object",
        "required": [
          "batch_id",
          "filename",
          "columns",
          "mapping",
          "total_rows",
          "counts",
          "rows",
          "target_fields"
        ],
        "title": "ImportPreview"
      },
      "ImportPreviewRow": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "values": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Values"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "duplicate",
              "invalid"
            ],
            "title": "Status"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of"
          },
          "duplicate_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Name"
          },
          "match_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Reason"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Issues"
          }
        },
        "type": "object",
        "required": [
          "index",
          "values",
          "status",
          "duplicate_of",
          "duplicate_name",
          "match_reason",
          "issues"
        ],
        "title": "ImportPreviewRow"
      },
      "ImportResult": {
        "properties": {
          "created": {
            "type": "integer",
            "title": "Created"
          },
          "merged": {
            "type": "integer",
            "title": "Merged"
          },
          "skipped_duplicates": {
            "type": "integer",
            "title": "Skipped Duplicates"
          },
          "invalid": {
            "type": "integer",
            "title": "Invalid"
          },
          "audit_job_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audit Job Id"
          }
        },
        "type": "object",
        "required": [
          "created",
          "merged",
          "skipped_duplicates",
          "invalid",
          "audit_job_id"
        ],
        "title": "ImportResult"
      },
...
      "LeadStatus": {
        "type": "string",
        "enum": [
          "NEW",
          "CALLED",
          "NO_ANSWER",
          "CALLBACK",
          "INTERESTED",
          "QUALIFIED",
          "PROPOSAL",
          "WON",
          "LOST",
          "DO_NOT_CONTACT"
        ],
        "title": "LeadStatus"
      },
...
      "OpportunityType": {
        "type": "string",
        "enum": [
          "NO_WEBSITE",
          "BROKEN_WEBSITE",
          "OUTDATED_WEBSITE",
          "WEBSITE_REDESIGN",
          "MOBILE_PROBLEM",
          "NO_BOOKING",
          "NO_CONTACT_CTA",
          "WEAK_CONVERSION",
          "SLOW_WEBSITE",
          "GOOGLE_PROFILE",
          "LOW_REVIEWS",
          "MISSING_BUSINESS_INFO"
        ],
        "title": "OpportunityType"
      },
...
      "Priority": {
        "type": "string",
        "enum": [
          "HOT",
          "WARM",
          "COLD"
        ],
        "title": "Priority"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}